Main to NHS

+++ RemovePackagingFactory.ps1
@@ -27,10 +27,19 @@ Set-Item Env:\SuppressAzurePowerShellBreakingChangeWarnings "true"  # Turns off 
 #endregion Setup
 
 #region Main
+if($isProd) { Write-Warning "Are you sure you want to delete the Packaging Environment?  OK to Continue?" -WarningAction Inquire }
+
+if(!($isProd) -and $RequireUserGroups) {
+    Remove-AzAdGroup -DisplayName $rbacOwner -ErrorAction Ignore  -Force -Verbose
+    Remove-AzAdGroup -DisplayName $rbacContributor -ErrorAction Ignore  -Force -Verbose
+    Remove-AzAdGroup -DisplayName $rbacReadOnly -ErrorAction Ignore  -Force -Verbose
+}
 Remove-AzResourceGroup -Name $RGNameDEV -Force -ErrorAction Ignore -Verbose
 Remove-AzResourceGroup -Name $RGNamePROD -Force -ErrorAction Ignore -Verbose
-Remove-AzResourceGroup -Name $RGNameDEVVNET -Force -ErrorAction Ignore -Verbose         # Dont want to do this is a Production Environment
-Remove-AzResourceGroup -Name $RGNamePRODVNET -Force -ErrorAction Ignore -Verbose        # Dont want to do this is a Production Environment
+if(!($isProd)) {
+    Remove-AzResourceGroup -Name $RGNameDEVVNET -Force -ErrorAction Ignore -Verbose         # Dont want to do this is a Production Environment
+    Remove-AzResourceGroup -Name $RGNamePRODVNET -Force -ErrorAction Ignore -Verbose        # Dont want to do this is a Production Environment
+}
 #endregion
